Converted most structure functionality to be compatible with C.
I have begun changing structures so that fields are accessed as a
positive offset from the start of the struct. This is beneficial
as this is the standard way to handle structs accross many different
languages, such as C. (Note: structs used to be implemented in reverse
order ie. they were reversed in memory and addressed from the end).

This includes two main changes:
	in AST.hs changes were made to changeVar and makeNewStack
	so that variables now point to the lowest byte in the type;
	conveniently, this simplified the code.

	in Compiler.hs most of the store and load functions for structs
	have been "reversed", with the notable exceptions of
	storeStructOff and loadStructOff, this means that we have all
	the previous functionality of structs *except* embedded
	accesses.

NOTE: we still need to implement storeStructOff etc.